extract arg parser code to separate file
